# openapi.yaml
openapi: 3.1.0
info:
  title: CivicLens API
  version: 1.0.0
  description: >
    AI-assisted public-service navigation API. Eligibility determinations are
    always produced by the deterministic rule engine (see
    docs/ai/eligibility-engine.md); the assistant endpoints are
    retrieval-grounded and always cite sources.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/otp/request:
    post:
      summary: Request an OTP for phone-based login/registration
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number]
              properties:
                phone_number: { type: string }
      responses:
        "202": { description: OTP dispatched }
        "429": { description: Rate limited, $ref: '#/components/responses/RateLimited' }
  /auth/otp/verify:
    post:
      summary: Verify OTP and issue tokens (registers account if new)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number, code]
              properties:
                phone_number: { type: string }
                code: { type: string }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
  /auth/register:
    post:
      summary: Register a new account with email+password (staff/admin or optional extension)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
  /auth/login:
    post:
      summary: Login with email+password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/refresh:
    post:
      summary: Exchange a refresh token for a new access token (alias of /auth/token/refresh)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/token/refresh:
    post:
      summary: Exchange a refresh token for a new access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      summary: Revoke the current refresh token (or all, with all=true)
      parameters:
        - in: query
          name: all
          schema: { type: boolean, default: false }
      responses:
        "204": { description: Logged out }

  # ---------- Citizen profile ----------
  /me:
    get:
      summary: Get the current citizen's profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CitizenProfile' }
    patch:
      summary: Update profile fields (progressive profiling)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CitizenProfileUpdate' }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CitizenProfile' }
        "422": { $ref: '#/components/responses/ValidationError' }
...
  /me/addresses:
    get:
      summary: List the current citizen's addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Address' }
    post:
      summary: Add an address
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Address' }
  /me/addresses/{address_id}:
    put:
      summary: Update address
      parameters:
        - name: address_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Updated }
    delete:
      summary: Delete address
      parameters:
        - name: address_id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Deleted }

  # ---------- Schemes ----------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
            field_errors:
              type: array
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: bearer }
        expires_in: { type: integer }
    CitizenProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        date_of_birth: { type: string, format: date, nullable: true }
        gender: { type: string, nullable: true }
        category: { type: string, nullable: true }
        occupation: { type: string, nullable: true }
        declared_annual_income: { type: number, nullable: true }
        disability_status: { type: boolean, nullable: true }
        family_size: { type: integer, nullable: true }
        profile_completeness: { type: number, description: "0-1, drives progressive profiling prompts" }
        current_version_no: { type: integer }
    CitizenProfileUpdate:
      type: object
      description: Any subset of CitizenProfile's editable fields
      additionalProperties: true
    Address:
      type: object
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [permanent, current] }
        state: { type: string }
        district: { type: string }
        pincode: { type: string }
        line1: { type: string }
    AddressInput:
      type: object
      required: [type, state, district, pincode, line1]
      properties:
        type: { type: string, enum: [permanent, current] }
        state: { type: string }
        district: { type: string }
        pincode: { type: string }
        line1: { type: string }
...